Fix Tcl 9 -Wall/-Wextra build warnings
Initialize the keyed-list Tcl_ObjType with TCL_OBJTYPE_V0 when
building against Tcl 9, where Tcl_ObjType has an explicit
version field.

Also use size-appropriate types in a few helper paths so builds
with -Wall -Wextra do not report signedness warnings for mode
and size comparisons.

This leaves the unused-parameter warnings alone, since those
mostly come from Tcl callback signatures and are better handled
separately.

+++ generic/tclXhandles.c
@@ -24,12 +24,12 @@
  * Variable set to contain the alignment factor (in bytes) for this machine.
  * It is set on the first table initialization.
  */
-static int entryAlignment = 0;
+static size_t entryAlignment = 0;
 
 /*
  * Rounded size of an entry header
  */
-static int entryHeaderSize = 0;
+static size_t entryHeaderSize = 0;
 
 /*
  * Marco to rounded up a size to be a multiple of (void *).  This is required
